AH-1781 WIP: More details on provenance
Additional detail in the provenance records added. All source files are now recorded in the provenance data, along with modification datetimes.

Added saving station data to a GeoJSON file so it can be directly imported into a file geodatabase. CRS is fixed to GDA2020 at this time.

Modified the category names in the configuration file to provide better clarity in provenance files